Generalize SdJwtVcVerifierFactory and extract platform code to separate files.
