+++ src/main/kotlin/eu/europa/ec/eudi/sdjwt/Types.kt
@@ -15,7 +15,6 @@
  */
 package eu.europa.ec.eudi.sdjwt
 
-import eu.europa.ec.eudi.sdjwt.HashAlgorithm.entries
 import kotlinx.serialization.json.JsonElement
 import kotlin.contracts.InvocationKind
 import kotlin.contracts.contract
